+++ tests/cli/test_commands.py
@@ -2,8 +2,6 @@
 # -*- coding: utf-8 -*-
 from flask_mail import Message
 from patientMatcher.cli.commands import cli
-from patientMatcher.parse.patient import mme_patient
-from patientMatcher.utils.ensembl_rest_client import requests
 
 
 def test_appname(mock_app):
@@ -37,4 +35,3 @@ def test_sendemail(mock_app, mock_mail):
     # Make sure that mock mail send method was called and mock email is sent
     assert mock_mail._send_was_called
     assert mock_mail._message
-    assert "Mail correctly sent" in result.output